gs-scheduler/global_scheduler2/policy_dockerfile/mostrequested/requirements.txt
Flask
redis
kafka-python
kubernetes
